feat: SAP app cleanup - move all component files into components folder [EXT-5652] (#8731)
* move appconfig into components folder and attempt to add paths

* cleanup import paths that weren't working

* add paths to resolve in vite config
